0-print_list: function that prints a list of integers, correction of output

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -6,9 +6,9 @@
  * @h: pointer to the head of the list
  * Return: count of nodes
  */
-unsigned long print_listint(const listint_t *h)
+size_t print_listint(const listint_t *h)
 {
-	unsigned long i = 0;
+	size_t i = 0;
 
 	while (h)
 	{
